requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
tqdm==4.66.1
nltk==3.8.1
scikit-learn==1.3.0
torch==2.0.1
transformers==4.34.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
streamlit==1.28.0
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
ipykernel==6.25.0